chem_qsar/hiv_classification/requirements.txt
torch_geometric
rdkit
torchmetrics
mordred
deepchem